Customize cert verify result
To include additional information that indicates this certificate
was verified by DNSSEC/DANE, we can modify cert verify result
but this will need lots of patching to wire it to the UI. Since
the X509 Certificate is already encoded and passed to the UI
in cert verify result, we update its pickle encoder to pass
the additional info.

This info doesn't change the meaning of the validation.
Only used to show "Secured by Handshake".